LayoutTests/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-parse-error-failure-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token '*'. import call expects exactly one argument.

Harness Error (FAIL), message = Error: assert_implements: Static import must be supported on module dedicated worker to run this test. undefined

NOTRUN Module worker construction for script with syntax error should dispatch an event named error.
NOTRUN Static import on module worker for script with syntax error should dispatch an event named error.
